chatbot/backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OpenAI and AI frameworks
openai>=1.6.0
# Note: MCP SDK may need to be installed from specific source
# mcp>=0.1.0

# Database and ORM
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
